refactor: extract useAmbientDemo to shared hook
